- added SampleModel project under misc folder; goal is to use the same model for all unit tests and sample/example projects - move model classes from RuleEngine.Tests project; added SampleModel ref - and updated unit tests to use the new models.
