+++ misc/SampleModel/CoOrdinate.cs
@@ -1,4 +1,4 @@
-﻿namespace RuleEngine.Tests.Model
+﻿namespace SampleModel
 {
     public class CoOrdinate
     {
