+++ src/RuleEngine.Tests/Rules/ConditionalRulesTests.cs
@@ -2,7 +2,7 @@
 using System.Linq.Expressions;
 using FluentAssertions;
 using RuleEngine.Rules;
-using RuleEngine.Tests.Model;
+using SampleModel;
 using Xunit;
 using Xunit.Abstractions;
