+++ src/RuleEngine.Tests/Rules/ExpressionActionRulesTests.cs
@@ -2,7 +2,7 @@
 using FluentAssertions;
 using RuleEngine.Rules;
 using RuleEngine.Tests.Fixture;
-using RuleEngine.Tests.Model;
+using SampleModel;
 using Xunit;
 using Xunit.Abstractions;
