+++ src/RuleEngine.Tests/Utils/ExressionExtensions/MethodCallExpressionTraceTests.cs
@@ -2,8 +2,8 @@
 using System.Linq.Expressions;
 using System.Text;
 using FluentAssertions;
-using RuleEngine.Tests.Model;
 using RuleEngine.Utils;
+using SampleModel;
 using Xunit;
 using Xunit.Abstractions;
